PLAT-66: Support traffic coming from API Platform.
This change adds an additional server configuration, which specifically listens for requests with a host matching the pattern:

    ^(?<name>.+)\.(?<environment>.+)\.tax\.service\.gov\.uk$

This represents API Platform traffic, destined for a protected microservice, which ingresses to the MDTP VPC from API Gateway.

NGINX then rewrites the request to the form `<name>.protected.mdtp` and proxies it to the appropriate microservice.

The proxied request is then intercepted by Envoy which proxies it, within the mesh, to an instance of the desired microservice.